Amazing refactoring stage 1. Add notification by updates.

+++ hr-core/src/main/java/com/digdes/services/DepartmentDataService.java
@@ -5,7 +5,11 @@ import com.digdes.dto.DepartmentResponseDto;
 import com.digdes.dto.EmployeeResponseDto;
 import com.digdes.exceptions.EntityDeleteException;
 import com.digdes.exceptions.EntityNotFoundException;
+import com.digdes.message.Message;
+import com.digdes.message.impl.EmailMessage;
 import com.digdes.models.*;
+import com.digdes.notifier.EmailNotifier;
+import com.digdes.notifier.Notifier;
 import com.digdes.repositories.DepartmentRepository;
 import com.digdes.repositories.DepartmentTypeRepository;
 import com.digdes.repositories.EmployeeRepository;
@@ -15,12 +19,13 @@ import org.springframework.data.domain.Example;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
-public class DepartmentDataService extends DataService{
+public class DepartmentDataService extends DataService<DepartmentDto, DepartmentResponseDto>{
 
     //todo: transactonal
     @Autowired
@@ -29,16 +34,21 @@ public class DepartmentDataService extends DataService{
     private EmployeeRepository employeeRepository;
     @Autowired
     private DepartmentTypeRepository typeRepository;
-
     @Autowired
     private UsersRepository usersRepository;
+
+    @Autowired
+    private Notifier notifier;
 
     @Transactional
     public DepartmentResponseDto create(DepartmentDto info) {
         try {
             Department department = mapDtoToDepartment(info);
             Optional<Department> existingDepartment = departmentRepository.findOne(Example.of(department));
-            return existingDepartment.map(this::mapDepartmentToResponseDto).orElseGet(() -> mapDepartmentToResponseDto(departmentRepository.save(department)));
+            Department newDepartment = existingDepartment.orElseGet(() -> (departmentRepository.save(department)));
+            if (department.getParent() != null)
+                notifier.sendMessage(getCreateMessage(newDepartment));
+            return mapDepartmentToResponseDto(newDepartment);
         }
         catch (EntityNotFoundException e){
             throw new EntityNotFoundException(e.getMessage());
@@ -50,10 +60,15 @@ public class DepartmentDataService extends DataService{
     public DepartmentResponseDto update(DepartmentDto info) {
         try {
             Department department = mapDtoToDepartment(info);
-            Optional<Department> updateDepartment = departmentRepository.findById(department.getId());
-            if (!updateDepartment.isPresent())
+            Optional<Department> existingDepartment = departmentRepository.findById(department.getId());
+            if (existingDepartment.isEmpty())
                 throw new EntityNotFoundException("Updated department is not found");
-            return mapDepartmentToResponseDto(departmentRepository.save(getUpdateDepartment(department, updateDepartment.get())));
+            Department updateDepartment = existingDepartment.get();
+            Department newDepartment = departmentRepository.save(getUpdateDepartment(department, updateDepartment));
+            for (Message message : getUpdateMessages(department, updateDepartment, newDepartment)) {
+                notifier.sendMessage(message);
+            }
+            return mapDepartmentToResponseDto(newDepartment);
         }
         catch (EntityNotFoundException e){
             throw new EntityNotFoundException(e.getMessage());
@@ -68,11 +83,18 @@ public class DepartmentDataService extends DataService{
                 throw new EntityDeleteException("Reference on this department in table employees (department_id)");
             if (departmentRepository.findAll(Example.of(new Department(department))).size() != 0)
                 throw new EntityDeleteException("Reference on this department in table department (parent_id)");
-            departmentRepository.delete(department);
-            return !departmentRepository.existsById(department.getId());
+            Optional<Department> oldDepartment = departmentRepository.findById(department.getId());
+            if (oldDepartment.isPresent()) {
+                departmentRepository.delete(department);
+                boolean result = !departmentRepository.existsById(department.getId());
+                if (department.getParent() != null)
+                    notifier.sendMessage(getDeleteMessage(department));
+                return result;
+            }
         } catch (EntityNotFoundException e) {
             throw new EntityNotFoundException(e.getMessage());
         }
+        return true;
     }
 
     @Transactional
@@ -89,13 +111,14 @@ public class DepartmentDataService extends DataService{
     }
 
     @Transactional(readOnly = true)
-    public DepartmentResponseDto get(Long id) {
+    public Optional<DepartmentResponseDto> get(Long id) {
         Optional<Department> department = departmentRepository.findById(id);
-        return department.map(this::mapDepartmentToResponseDto).orElseGet(DepartmentResponseDto::new);
+        return department.map(this::mapDepartmentToResponseDto);
     }
 
 
     @Transactional(readOnly = true)
+    @Override
     public List<DepartmentResponseDto> getAll() {
         return departmentRepository.findAll()
                 .stream().map(this::mapDepartmentToResponseDto)
@@ -187,4 +210,78 @@ public class DepartmentDataService extends DataService{
             updateDepartment.setEmployees(info.getEmployees());
         return updateDepartment;
     }
+
+    //todo: create department --- heads of parent if not null
+    private Message getCreateMessage(Department department){
+        String[] receivers = getReceivers(department, true).toArray(String[]::new);
+        Message message = new EmailMessage();
+        message.setFrom(EmailNotifier.FROM_EMAIL);
+        message.setTo(receivers);
+        message.setSubject("Subdivision created in your department");
+        message.setBody(String.format("Hello!\n Subdivision \"%s\" created in your department.", department.getName()));
+        return message;
+    }
+
+    //todo: update department --
+    private List<Message> getUpdateMessages(Department updateInfo, Department updateDepartment, Department newDepartment) {
+        //todo: check empty moders
+        List<Message> messages = new ArrayList<>();
+        // todo: name, type, parent, emplyees --- all employees
+        if (updateInfo.getName() != null || updateInfo.getType() != null
+                || updateInfo.getParent() != null) {
+            String[] receivers = getReceivers(updateInfo, null).toArray(String[]::new);
+            messages.add(getAllEmployeesMessage(receivers, newDepartment));
+        }
+        // todo: moderator --- old moderator, new if not null
+        if (updateInfo.getModerator() != null)
+            messages.add(getUpdateModeratorMessage(updateInfo, newDepartment, newDepartment));
+        else if (updateDepartment.getModerator() != null)
+            messages.add(getUpdateModeratorMessage(updateInfo, newDepartment, newDepartment));
+        return messages;
+    }
+
+    private Message getUpdateModeratorMessage(Department updateInfo, Department updateDepartment, Department newDepartment){
+        List<String> receivers = new ArrayList<>();
+        if (updateInfo.getModerator() != null)
+            employeeRepository
+                    .findEmployeeByUser(updateInfo.getModerator())
+                    .map(Employee::getEmail)
+                    .ifPresent(receivers::add);
+        if (updateDepartment.getModerator() != null)
+            employeeRepository
+                    .findEmployeeByUser(updateDepartment.getModerator())
+                    .map(Employee::getEmail)
+                    .ifPresent(receivers::add);
+        return getUpdateModeratorMessage(receivers.toArray(String[]::new), newDepartment);
+    }
+
+    private Message getAllEmployeesMessage(String[] receivers, Department newDepartment){
+        Message message = new EmailMessage();
+        message.setFrom(EmailNotifier.FROM_EMAIL);
+        message.setTo(receivers);
+        message.setSubject("Change data of department");
+        message.setBody(String.format("Hello!\n New data of department is %s.", newDepartment.toString()));
+        return message;
+    }
+
+    private Message getUpdateModeratorMessage(String[] receivers, Department newDepartment){
+        Message message = new EmailMessage();
+        message.setFrom(EmailNotifier.FROM_EMAIL);
+        message.setTo(receivers);
+        message.setSubject("Change moderator of department");
+        message.setBody(String.format("Hello!\n New moderator of department \"%s\" is \"%s\".", newDepartment.getName(), newDepartment.getModerator().getUsername()));
+        return message;
+    }
+
+    //todo: delete department --- heads of parent if not null
+    private Message getDeleteMessage(Department oldDepartment){
+        String[] receivers = getReceivers(oldDepartment, true).toArray(String[]::new);
+        Message message = new EmailMessage();
+        message.setFrom(EmailNotifier.FROM_EMAIL);
+        message.setTo(receivers);
+        message.setSubject("Subdivision deleted from your department");
+        message.setBody(String.format("Hello!\n Subdivision \"%s\" deleted from your department.", oldDepartment.getName()));
+        return message;
+    }
+
 }
